+++ app/Http/Controllers/SubscriptionController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\ModificationRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Subscription;
@@ -32,7 +33,7 @@ class SubscriptionController extends Controller
             $subscriptionsQuery->where('user_id', auth()->user()->id);
         }
 
-        $subscriptions = $subscriptionsQuery->paginate(10);
+        $subscriptions = $subscriptionsQuery->orderBy('created_at', 'desc')->paginate(10);
 
         return view('subscriptions.index', compact('subscriptions', 'activeLawyers'));
     }
@@ -111,13 +112,33 @@ class SubscriptionController extends Controller
         ]);
 
         $subscription = Subscription::findOrFail($id);
-        $subscription->user_id = $request->user_id;
-        $subscription->subscription_type = $request->subscription_type;
-        $subscription->start_date = $request->start_date;
-        $subscription->end_date = $request->end_date;
-        $subscription->save();
 
-        return redirect()->route('subscriptions')->with('success', 'Subscription updated successfully.');
+        if ($subscription) {
+            if (auth()->user()->hasRole('president')) {
+
+                $subscription->user_id = $request->user_id;
+                $subscription->subscription_type = $request->subscription_type;
+                $subscription->start_date = $request->start_date;
+                $subscription->end_date = $request->end_date;
+                $subscription->save();
+
+                return redirect()->route('subscriptions')->with('success', 'Subscription updated successfully.');
+            } else {
+
+                $changes = $request->except(['_token', '_method']);
+                $this->submitChangeRequest([
+                    "table_name" => 'subscriptions',
+                    "record_id" => $subscription->id,
+                    "changes" => $changes,
+                    "action" => ModificationRequest::REQUEST_TYPE_UPDATE,
+                    "requested_by" => Auth::id(),
+                ]);
+
+                return redirect()->route('subscriptions')->with('success', 'Subscription updated request submitted successfully.');
+            }
+        }
+
+        return redirect()->route('subscriptions')->with('error', 'Something went wrong.');
     }
 
     public function destroy($id)
@@ -125,13 +146,28 @@ class SubscriptionController extends Controller
         // Find the user by ID
         $subscription = Subscription::findOrFail($id);
 
-        // Set the deleted_by field with the authenticated subs$subscription's ID
-        $subscription->deleted_by = Auth::id();
-        $subscription->save(); // Save the subs$subscription to update the deleted_by field
+        if ($subscription) {
+            if (auth()->user()->hasRole('president')) {
+                // Set the deleted_by field with the authenticated subs$subscription's ID
+                $subscription->deleted_by = Auth::id();
+                $subscription->save(); // Save the subs$subscription to update the deleted_by field
 
-        // Soft delete the subs$subscription
-        $subscription->delete();
+                // Soft delete the subs$subscription
+                $subscription->delete();
 
-        return redirect()->route('subscriptions')->with('success', 'Subscription deleted successfully!');
+                return redirect()->route('subscriptions')->with('success', 'Subscription deleted successfully!');
+            } else {
+                $this->submitChangeRequest([
+                    "table_name" => 'subscriptions',
+                    "record_id" => $subscription->id,
+                    "action" => ModificationRequest::REQUEST_TYPE_DELETE,
+                    "requested_by" => Auth::id(),
+                ]);
+
+                return redirect()->route('subscriptions')->with('success', 'Subscription updated request submitted successfully.');
+            }
+        }
+
+        return redirect()->route('subscriptions')->with('error', 'Something went wrong.');
     }
 }
